some changes

+++ database/migrations/2024_02_10_075054_create_cash_outs_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->double('cashout');
             $table->integer('status');
-            $table->integer('number');
+            $table->double('number');
             $table
                 ->foreignIdFor(User::class)
                 ->constrained()
